third_party/blink/web_tests/fast/forms/number/number-reject-invalid-expected.txt

Arabic number input should accept ASCII digits and Arabic digits, and reject others.
PASS displayValueForKeyInput(arabicInput, "123.4") is "123.4"
PASS displayValueForKeyInput(arabicInput, "1.23E+19") is "1.23E+19"
PASS displayValueForKeyInput(arabicInput, "1.23e-1") is "1.23e-1"
PASS displayValueForKeyInput(arabicInput, "١٩٠") is "١٩٠"
PASS displayValueForKeyInput(arabicInput, "acdef") is "e"


French number input should accept ASCII digits, comma, and full stop.
PASS displayValueForKeyInput(frenchInput, "1234.56") is "1234.56"
FAIL displayValueForKeyInput(frenchInput, "1234,56") should be 1234.56. Was 1234,56.


English number input should accept ASCII digits and full stop, and no comma.
PASS displayValueForKeyInput(englishInput, "1234.56") is "1234.56"
PASS displayValueForKeyInput(englishInput, "-1234,56") is "-123456"
PASS displayValueForKeyInput(englishInput, " 1234.56 ") is "1234.56"
PASS displayValueForKeyInput(englishInput, "123,456,789E+10") is "123456789E+10"
PASS displayValueForKeyInput(englishInput, ".", "1|e2") is "1.e2"
PASS displayValueForKeyInput(englishInput, ".", "1e2|") is "1e2"
PASS displayValueForKeyInput(englishInput, ".", "|-12") is "-12"
PASS displayValueForKeyInput(englishInput, ".", "|1e-12") is ".1e-12"
PASS displayValueForKeyInput(englishInput, "e", "12|34") is "12e34"
PASS displayValueForKeyInput(englishInput, "e", "12|3e4") is "123e4"
PASS displayValueForKeyInput(englishInput, "e", "123|.4") is "123.4"
PASS displayValueForKeyInput(englishInput, "e", "12.3|4") is "12.3e4"
PASS displayValueForKeyInput(englishInput, "+", "12|34") is "12+34"
PASS displayValueForKeyInput(englishInput, "+", "-1|2e-34") is "-12e-34"
PASS displayValueForKeyInput(englishInput, "+", "|1234") is "+1234"
PASS displayValueForKeyInput(englishInput, "-", "123e|4") is "123e-4"
PASS displayValueForKeyInput(englishInput, "-", "1|23e4") is "123e4"
PASS displayValueForKeyInput(englishInput, "-", "123e4|") is "123e4"
PASS displayValueForKeyInput(englishInput, "9", "|-1") is "-1"
PASS displayValueForKeyInput(englishInput, "9", "-|1") is "-91"
PASS displayValueForKeyInput(englishInput, "9", "1e|+2") is "1e+2"
PASS displayValueForKeyInput(englishInput, "9", "1e+|2") is "1e+92"
PASS displayValueForKeyInput(englishInput, " abcdef ") is "e"
PASS displayValueForKeyInput(englishInput, "+1-2") is "+1-2"
PASS displayValueForKeyInput(englishInput, "+1-2+2-3") is "+1-223"
PASS displayValueForKeyInput(englishInput, "0-123-123+123") is "0-123-123123"
PASS displayValueForKeyInput(englishInput, "10e123123e1231233e") is "10e1231231231233"
PASS displayValueForKeyInput(englishInput, "1e2eee") is "1e2"
PASS displayValueForKeyInput(englishInput, "1e1e1e1e") is "1e111"
PASS successfullyParsed is true

TEST COMPLETE

  
